Test ORK converters
