+++ MindsetTests/Info.plist
@@ -18,5 +18,7 @@
 	<string>1.0</string>
 	<key>CFBundleVersion</key>
 	<string>1</string>
+	<key>LSApplicationCategoryType</key>
+	<string>ORKConverters</string>
 </dict>
 </plist>
